pyproject.toml
[project]
name = "aichef"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "chromadb>=1.3.5",
    "fastapi>=0.123.0",
    "langchain>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.1.0",
    "langchain-openai>=1.1.0",
    "numpy>=2.1",
    "posthog<3.5.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "torch>=2.9.1",
    "uvicorn>=0.38.0",
]
